fixed clip type in multistart_optimizer.py

+++ qiskit/optimization/algorithms/multistart_optimizer.py
@@ -121,7 +121,7 @@ class MultiStartOptimizer(OptimizationAlgorithm, ABC):
         return self._clip
 
     @clip.setter
-    def clip(self, clip: int) -> None:
+    def clip(self, clip: float) -> None:
         """Sets the clip value.
 
         Args:
